requirements.txt
fastapi
sse-starlette
uvicorn[standard]
python-dotenv
psycopg[binary]